# Use print functions if there is no format